+++ src/pages/admin/Proveedores.tsx
@@ -1,4 +1,4 @@
-import { useState, useEffect  } from 'react';
+import { useState, useEffect, useContext  } from 'react';
 import { AddOutlined, CategoryOutlined } from '@mui/icons-material';
 import { Box, Button, Chip, Grid } from '@mui/material'
 import { DataGrid, GridColDef, GridRenderCellParams, GridValueGetterParams } from '@mui/x-data-grid';
@@ -8,9 +8,11 @@ import { NavLink } from 'react-router-dom';
 import { AdminLayoutMenuList } from '../../components/layouts'
 import { ISupplier  } from '../../interfaces';
 import { stutzApi } from '../../../api';
+import { AuthContext } from '../../../context';
 
 
 export const Proveedores = () => {
+const { user } = useContext(AuthContext);
 
 const columns:GridColDef[] = [
     { field: 'codSup', headerName: 'Codigo' },
@@ -36,6 +38,7 @@ const columns:GridColDef[] = [
             field: 'check',
             headerName: 'Acción',
             renderCell: ({ row }: GridValueGetterParams | GridRenderCellParams ) => {
+                if (user?.role !== 'admin') return null;
                 return (
                         <Chip variant='outlined' label="Eliminar" color="error"
                         onClick={() => deleteHandler(row.id)}
